style: 🎨 landing page top bar
