+++ src/components/layout/TradingHours/TradingHours.view.jsx
@@ -1,9 +1,39 @@
 import React, { forwardRef } from 'react';
-import { Grid } from '@material-ui/core';
+import { Button, Grid, withStyles } from '@material-ui/core';
 import { useStyles } from '../../../styles/MainStyles';
 import BellNotifications from '../topbarParts/BellNotifications/BellNotifications';
 import UserIcon from '../topbarParts/UserIcon/UserIcon';
 import TradingHourUnit from '../topbarParts/TradingHourUnit/TradingHourUnit';
+
+const StyledButton = withStyles(() => ({
+	root: {
+		'width': '185px',
+		'height': '30px',
+		'textTransform': 'none',
+		'fontWeight': 400,
+		'backgroundColor': '#1C67FF',
+		'borderRadius': 21,
+		'color': '#F2F2F2',
+		'&:hover': {
+			backgroundColor: '#1c67ffb3',
+		},
+	},
+}))(Button);
+
+const LoginButton = withStyles(() => ({
+	root: {
+		'width': '122px',
+		'height': '30px',
+		'textTransform': 'none',
+		'fontWeight': 400,
+		'backgroundColor': '#ffffff',
+		'borderRadius': 21,
+		'color': '#000000',
+		'&:hover': {
+			backgroundColor: '#ffff',
+		},
+	},
+}))(Button);
 
 const TradingHoursView = forwardRef((props, ref) => {
 	const classes = useStyles();
@@ -23,26 +53,41 @@ const TradingHoursView = forwardRef((props, ref) => {
 				justifyContent="flex-end"
 				className={classes.userBarWrapper}
 			>
-				<Grid item xs={1}>
-					<BellNotifications
-						handleToggle={props.handleToggle}
-						notifications={props.notifications}
-						openNotification={props.openNotification}
-						setOpenNotification={props.setOpenNotification}
-						handleListKeyDown={props.handleListKeyDown}
-						handleClose={props.handleClose}
-					/>
-				</Grid>
-				<Grid item className={classes.userIconWrapper} xs={3}>
-					<UserIcon
-						ref={ref}
-						handleToggle={props.handleToggle}
-						userType={props.userType}
-						handleClose={props.handleClose}
-						setOpen={props.setOpen}
-						open={props.open}
-					/>
-				</Grid>
+				{props.isAuthenticated ? (
+					<>
+						<Grid item xs={1}>
+							<BellNotifications
+								handleToggle={props.handleToggle}
+								notifications={props.notifications}
+								openNotification={props.openNotification}
+								setOpenNotification={props.setOpenNotification}
+								handleListKeyDown={props.handleListKeyDown}
+								handleClose={props.handleClose}
+							/>
+						</Grid>
+						<Grid item className={classes.userIconWrapper} xs={3}>
+							<UserIcon
+								ref={ref}
+								handleToggle={props.handleToggle}
+								userType={props.userType}
+								handleClose={props.handleClose}
+								setOpen={props.setOpen}
+								open={props.open}
+							/>
+						</Grid>
+					</>
+				) : (
+					<Grid item style={{ paddingRight: 80 }}>
+						<Grid container justifyContent="space-between">
+							<Grid item style={{ paddingRight: 20 }}>
+								<StyledButton>Become a Costumer</StyledButton>
+							</Grid>
+							<Grid item>
+								<LoginButton>Login</LoginButton>
+							</Grid>
+						</Grid>
+					</Grid>
+				)}
 			</Grid>
 		</Grid>
 	);
